pc99: Add CONFIG_PC99_TSC_FREQUENCY config option
This option allows a static definition of the TSC frequency. This allows
a kernel to use a known value if the platform doesn't provide access to
an appropriate MSR for reading it during startup. If this config is set,
the kernel won't try and estimate the TSC frequency by running the PIT
for 200ms during startup.

Signed-off-by: Kent McLeod <kent@kry10.com>